Some Improvements for CXF RS Configuration

+++ geoplatform-services/geoplatform-ws-core/geoplatform-ws-core-api/src/main/java/org/geosdi/geoplatform/responce/ServerDTO.java
@@ -167,7 +167,7 @@ public class ServerDTO implements Serializable {
      */
     @Override
     public String toString() {
-        return "ServerDTO{" + "id=" + id.longValue()
+        return "ServerDTO{" + "id=" + id
                 + ", serverUrl=" + serverUrl
                 + ", alias=" + alias
                 + ", name=" + name
